Rearraged new code
